+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/service/PurchaseOrderLineServiceImpl.java
@@ -187,7 +187,7 @@ public class PurchaseOrderLineServiceImpl implements PurchaseOrderLineService {
 
 		Product product = purchaseOrderLine.getProduct();
 		
-		if(product == null)  {  return BigDecimal.ZERO;  }
+		if(product == null || !product.getSellable())  {  return BigDecimal.ZERO;  }
 		
 		TaxLine saleTaxLine = accountManagementService.getTaxLine(
 				purchaseOrder.getOrderDate(), purchaseOrderLine.getProduct(), purchaseOrder.getCompany(), purchaseOrder.getSupplierPartner().getFiscalPosition(), false);
@@ -202,7 +202,7 @@ public class PurchaseOrderLineServiceImpl implements PurchaseOrderLineService {
 	@Override
 	public BigDecimal getSalePrice(PurchaseOrder purchaseOrder, Product product, BigDecimal price) throws AxelorException  {
 
-    if(product == null)  {  return BigDecimal.ZERO;  }
+    if(product == null || !product.getSellable())  {  return BigDecimal.ZERO;  }
 	  
 		TaxLine saleTaxLine = accountManagementService.getTaxLine(
 				purchaseOrder.getOrderDate(), product, purchaseOrder.getCompany(), purchaseOrder.getSupplierPartner().getFiscalPosition(), false);